#2005: extract timer-wheel sweeps/eviction into session/expire.rs
Pure code-motion split of the conntrack fast path. Move the bucketed
timer-wheel GC machinery off SessionTable's monolithic impl in
session/mod.rs into a focused session/expire.rs submodule:
last_pop_stats, wheel_observe, push_to_wheel, expire_stale_entries, and
expire_stale. Bodies are byte-for-byte identical; only the file
boundary changed.

Visibility: push_to_wheel was a file-private `fn` and is called across
the module boundary by the in-place-refresh contract that stays in
mod.rs (update_session / refresh_for_ha_transition / the lookup paths)
plus the (forthcoming) install/lookup submodules and the test oracle.
Widened to the absolute `pub(in crate::session)` so its scope is
preserved (visible to the whole session module tree), not narrowed to
expire.rs's file scope. wheel_observe stays file-private to expire.rs
(only its sibling methods there call it).

The SessionWheel/WheelEntry types and the wheel constants stay in
session/wheel.rs; expire.rs imports them via `super::wheel::{...}`.
mod.rs's now-unused `use wheel::{...}` glob is trimmed to SessionWheel
(the only wheel symbol mod.rs still names); the GC timer tests that
relied on the glob to reach WHEEL_TICK_NS/WHEEL_BUCKETS/FAR_FUTURE_OFFSET
via `super::*` now import them explicitly from super::wheel (same
symbols, no behavior change). The `mod expire;` declaration is placed
after the debug_log! macro so the macro is in textual scope for the
moved expire_stale_entries body.

#1752/#1855 in-place-refresh contract (single-writer &mut self, eager
#964 cleanup, secondary-index re-assert) is untouched and remains in
mod.rs.

Validation: cargo build --release clean; cargo test --release session::
78 session tests pass (incl. the timer-wheel K-bound/classification
tests and the in-place-refresh fused-diff oracle suite).
